# specs/002-rl-trading-agent/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: RL Trading Agent API
  version: 0.1.0
  description: API contracts for operating the RL trading agent on BingX gold perpetual pairs.
servers:
  - url: https://api.example.invalid
security:
  - bearerAuth: []
paths:
...
  /agents/{agentId}/source-policies:
    get:
      summary: List source gating policies
      parameters:
        - $ref: "#/components/parameters/agentId"
      responses:
        "200":
          description: Source policies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SourcePolicy"
    patch:
      summary: Update source gating policies
      parameters:
        - $ref: "#/components/parameters/agentId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: "#/components/schemas/SourcePolicy"
      responses:
        "200":
          description: Source policies updated
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SourcePolicy"
...
  /data-sources/status:
    get:
      summary: Get data source freshness status
      parameters:
        - in: query
          name: pair
          schema:
            $ref: "#/components/schemas/TradingPair"
      responses:
        "200":
          description: Source status list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DataSourceStatus"
  /data-sources/config:
    patch:
      summary: Enable or disable data sources and thresholds
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DataSourceConfig"
      responses:
        "200":
          description: Updated data source configuration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataSourceConfig"
  /data-sources/runs:
    get:
      summary: List data source ingestion runs
      parameters:
        - in: query
          name: sourceType
          schema:
            type: string
        - in: query
          name: pair
          schema:
            $ref: "#/components/schemas/TradingPair"
      responses:
        "200":
          description: Data source runs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DataSourceRun"
  /data-sources/backfill:
    post:
      summary: Trigger data source backfill
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DataSourceBackfillRequest"
      responses:
        "202":
          description: Backfill accepted
...
  /risk-limits:
    get:
      summary: List risk limit sets
      responses:
        "200":
          description: Risk limits
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RiskLimitSet"
    post:
      summary: Create risk limit set
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RiskLimitSet"
      responses:
        "201":
          description: Risk limit set created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RiskLimitSet"
  /risk-limits/{riskLimitId}:
    patch:
      summary: Update risk limit set
      parameters:
        - $ref: "#/components/parameters/riskLimitId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RiskLimitSet"
      responses:
        "200":
          description: Risk limit set updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RiskLimitSet"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    agentId:
      name: agentId
      in: path
      required: true
      schema:
        type: string
...
    riskLimitId:
      name: riskLimitId
      in: path
      required: true
      schema:
        type: string
...
  schemas:
    TradingPair:
      type: string
      enum: [Gold-USDT, XAUTUSDT, PAXGUSDT]
...
    DataSourceStatus:
      type: object
      properties:
        sourceType:
          type: string
          enum:
            - bingx_candles
            - bingx_orderbook
            - bingx_trades
            - bingx_funding
            - bingx_open_interest
            - bingx_mark_price
            - bingx_index_price
            - bingx_ticker
            - ideas
            - signals
            - news
            - ocr_text
            - trades
        pair:
          $ref: "#/components/schemas/TradingPair"
        status:
          type: string
          enum: [ok, stale, unavailable]
        lastSeenAt:
          type: string
          format: date-time
        freshnessThresholdSeconds:
          type: integer
        coveragePct:
          type: number
          nullable: true
        missingFieldsCount:
          type: integer
          nullable: true
        parseConfidence:
          type: number
          nullable: true
      required: [sourceType, status]
    DataSourceConfig:
      type: object
      properties:
        sources:
          type: array
          items:
            type: object
            properties:
              sourceType:
                type: string
                enum:
                  - bingx_candles
                  - bingx_orderbook
                  - bingx_trades
                  - bingx_funding
                  - bingx_open_interest
                  - bingx_mark_price
                  - bingx_index_price
                  - bingx_ticker
                  - ideas
                  - signals
                  - news
                  - ocr_text
                  - trades
              enabled:
                type: boolean
              freshnessThresholdSeconds:
                type: integer
              minCoveragePct:
                type: number
              maxMissingFields:
                type: integer
              minParseConfidence:
                type: number
            required: [sourceType, enabled]
      required: [sources]
...
    DataSourceRun:
      type: object
      properties:
        id:
          type: string
        sourceType:
          type: string
        pair:
          $ref: "#/components/schemas/TradingPair"
        status:
          type: string
          enum: [running, succeeded, failed]
        startedAt:
          type: string
          format: date-time
        finishedAt:
          type: string
          format: date-time
          nullable: true
        newCount:
          type: integer
        updatedCount:
          type: integer
        errorCount:
          type: integer
        errorSummary:
          type: string
          nullable: true
    DataSourceBackfillRequest:
      type: object
      properties:
        sourceType:
          type: string
        pair:
          $ref: "#/components/schemas/TradingPair"
        start:
          type: string
          format: date-time
        end:
          type: string
          format: date-time
      required: [sourceType, pair, start, end]
    SourcePolicy:
      type: object
      properties:
        sourceId:
          type: string
        allowTrading:
          type: boolean
        minConfidenceScore:
          type: number
...
    RiskLimitSet:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        maxPositionSize:
          type: number
        leverageCap:
          type: number
        maxDailyLoss:
          type: number
        maxDrawdown:
          type: number
        maxOpenPositions:
          type: integer
      required:
        - name
        - maxPositionSize
        - leverageCap
        - maxDailyLoss
        - maxDrawdown
        - maxOpenPositions
...